Work towards splitting out a separate particle mapping tool
